requirements.txt
# Core AI / Agent framework
crewai==0.80.0
crewai-tools==0.17.0

# LLM provider
anthropic==0.40.0

# Vector store / RAG
chromadb>=1.5.0
# sentence-transformers requires PyTorch>=2.4; ChromaDB's DefaultEmbeddingFunction
# (all-MiniLM-L6-v2 via ONNX) is used instead — no direct torch dependency needed.

# Data processing
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.5.2

# UI
streamlit==1.40.0
plotly==5.24.1
altair==5.5.0

# Utilities
python-dotenv==1.0.1
pydantic==2.10.3
tenacity==9.0.0
loguru==0.7.3
tqdm==4.67.1

# Dev / testing
pytest==8.3.4
pytest-asyncio==0.24.0
